% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyCitation.R
\name{applyCitations}
\alias{applyCitations}
\title{Apply to citations}
\usage{
applyCitations(
  pmids,
  fun,
  fetchSettings,
  connectionDetails,
  databaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  ...
)
}
\arguments{
\item{pmids}{A vector of PMIDs.}

\item{fun}{The function to call per citation. This function should have exactly 1 argument,
a data frame containing one row with the citation. This data frame will at least
have a \code{pmid} column, and more depending on the \code{fetchSettings} argument.}

\item{fetchSettings}{An object of type \code{fetchSettings} as created using \code{\link[=createFetchSettings]{createFetchSettings()}}.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using \code{DatabaseConnector::createConnectionDetails()}.}

\item{databaseSchema}{The database schema containing the MEDLINE database.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp
tables, provide a schema with write privileges where temp tables can be created.}

\item{...}{Additional arguments to pass to fun.}
}
\value{
Returns the list of results from the calls to \code{fun}.
}
\description{
Apply a function to each citation in a set.
}
