% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractEstimates.R
\name{computeLogRrAndSeLogRr}
\alias{computeLogRrAndSeLogRr}
\title{Compute logRr and seLogRr for estimates}
\usage{
computeLogRrAndSeLogRr(estimates)
}
\arguments{
\item{estimates}{A tibble containing effect-size estimates as extracted using the
\code{\link[=extractEstimates]{extractEstimates()}} function.}
}
\value{
A tibble containing effect-size estimates, with the logRr and seLogRr columns added.
}
\description{
Compute logRr and seLogRr for estimates
}
