% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyCitation.R
\name{createFetchSettings}
\alias{createFetchSettings}
\title{Create fetch settings}
\usage{
createFetchSettings(title = FALSE, abstract = FALSE)
}
\arguments{
\item{title}{Should the citation title be extracted? If TRUE, the 'title' column will be available.}

\item{abstract}{Should the abstract title be extracted? If TRUE, the 'abstact' column will be available.}
}
\value{
Returns an object of type \code{fetchSettings}.
}
\description{
Creates fetch settings to be used in the \code{\link[=applyCitations]{applyCitations()}} function.
}
