% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractEstimates.R
\name{extractEstimates}
\alias{extractEstimates}
\title{Extract effect-size estimates from text}
\usage{
extractEstimates(abstract)
}
\arguments{
\item{abstract}{The text from which to extract the estimates.}
}
\value{
A tibble with one row per estimate.
}
\description{
Extract effect-size estimates from text
}
