% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryOnlinePubMed.R
\name{getObservationalResearchQuery}
\alias{getObservationalResearchQuery}
\title{Get Observational research query.}
\usage{
getObservationalResearchQuery()
}
\value{
A query for finding all observational studies in PubMed.
}
\description{
Get Observational research query.
}
\references{
Schuemie MJ, Ryan PB, Hripcsak G, Madigan D, Suchard MA. Improving reproducibility by using high-throughput observational studies with empirical calibration. Philos Trans A Math Phys Eng Sci. 2018 Sep 13;376(2128):20170356. doi: 10.1098/rsta.2017.0356. PMID: 30082302; PMCID: PMC6107542.
}
