% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractEstimates.R
\name{jitterEstimates}
\alias{jitterEstimates}
\title{Jitter estimates}
\usage{
jitterEstimates(estimates)
}
\arguments{
\item{estimates}{A tibble containing effect-size estimates as extracted using the
\code{\link[=extractEstimates]{extractEstimates()}} function.}
}
\value{
A tibble containing effect-size estimates, with noise added.
}
\description{
Adds random noise to effect-size estimates such that the estimated statistics still
round to the numbers extracted from the text.
}
