% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedlineAnalysis.R
\docType{package}
\name{MedlineAnalysis-package}
\alias{MedlineAnalysis}
\alias{MedlineAnalysis-package}
\title{MedlineAnalysis: Analyse a Local MEDLINE Dabatabase}
\description{
Analyse a MEDLINE database generated using ohdsi/MedlineXmlToDatabase.
}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

}
\keyword{internal}
