% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractEstimates.R
\name{plotEstimates}
\alias{plotEstimates}
\title{Plot effect-size estimates}
\usage{
plotEstimates(estimates, jitter = TRUE, fileName = NULL)
}
\arguments{
\item{estimates}{A tibble containing effect-size estimates as extracted using the
\code{\link[=extractEstimates]{extractEstimates()}} function.}

\item{jitter}{Should the estimates be jittered when plotting?}

\item{fileName}{If provided, the plot will be saved to this file.}
}
\value{
Returns a ggplot plot.
}
\description{
Plot effect-size estimates
}
