% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryOnlinePubMed.R
\name{queryOnlinePubmed}
\alias{queryOnlinePubmed}
\title{Query online PubMed}
\usage{
queryOnlinePubmed(query)
}
\arguments{
\item{query}{A string representing a PubMed query.}
}
\value{
A vector of PMIDs.
}
\description{
Query online PubMed
}
